requirements.txt
langchain
langchain-community
pydantic
requests
python-dotenv